refactor: rewrite tests in ts
