+++ src/index.ts
@@ -1,5 +1,5 @@
-import createError from "http-errors";
-import { Request, Response, NextFunction } from "express";
+import { NextFunction, Request, Response } from "express";
+import * as createError from "http-errors";
 
 type validationObjectType = {
     location: "body" | "query" | "params";
